Define initial idea and API behind pygpib
The API is modeled on the following presumptions:
  * GPIB interfaces are enumerable
  * A GPIB interface may communicate with several instruments

The main point is that a gpib.Interface() is not low-level library for
micromanaging the GPIB bus. It doesn't expose functionality to control
GPIB handshake lines, or transition through the GPIB state machine.
Its focus is providing a pipe to an instrument.

Most USB adapters have high level commands for addressing instruments.
They handle GPIB details in firmware. On these adapters, we can
achieve communication with much fewer USB transactions than linux-gpib,
deceasing latency and increasing throughput.

These are the design goals.