gpib.py: Update API to allow GPIB adapters to be enumerated
The model is simple. Several interface drivers may exist. Each driver
knows how to scan and enumerate its devices. The top level
gpib.list_devices() simply puts together the lists of devices from
each driver.

A list of all adapters was chosen, instead of generators or other
representations because we expect to only ever have a very limited
number of adapters connected to a system.
This is subject to change.