gpib.py: Allow Instruments to have a read timeout
GPIB instruments may take a while to respond to queries. The exact
amount is dependent on instrument,a nd instrument state. For example,
H3457A will responf to an "ID?" query within 300 ms, on most ranges,
but may take up to 2 seconds on AC voltage.

Each instrument needs to be able to have its own timeout. While it is
possible to permit a different timeout for each read transaction, it
would lead to a messy API. Instead, make the read timeout value a part
of an isnterument's bus config. Device drivers are expected to honor
the timeout value.